+++ tools/cli/commands.js
@@ -344,7 +344,7 @@ function doRunCommand(options) {
 
   let webArchs = ['web.browser'];
   let cordovaRunner;
-  if (!_.isEmpty(runTargets)) {
+  if (!_.isEmpty(runTargets) || options['mobile-server']) {
     main.captureAndExit('', 'preparing Cordova project', () => {
       const cordovaProject = new CordovaProject(projectContext, {
         settingsFile: options.settings,
@@ -1145,7 +1145,6 @@ to this command.`);
   } else {
     // remote mode
     var site = qualifySitename(options.args[0]);
-    config.printUniverseBanner();
 
     mongoUrl = deploy.temporaryMongoUrl(site);
     usedMeteorAccount = true;
@@ -1240,7 +1239,6 @@ main.registerCommand({
   catalogRefresh: new catalog.Refresh.Never()
 }, function (options, {rawOptions}) {
   var site = options.args[0];
-  config.printUniverseBanner();
 
   if (options.delete) {
     return deploy.deleteApp(site);
@@ -1358,7 +1356,6 @@ main.registerCommand({
     return 1;
   }
 
-  config.printUniverseBanner();
   auth.pollForRegistrationCompletion();
   var site = qualifySitename(options.args[0]);
 
@@ -1390,7 +1387,6 @@ main.registerCommand({
   maxArgs: 1,
   catalogRefresh: new catalog.Refresh.Never()
 }, function (options) {
-  config.printUniverseBanner();
   auth.pollForRegistrationCompletion();
   var site = qualifySitename(options.args[0]);
 
@@ -1671,7 +1667,12 @@ function doTestCommand(options) {
 
   return runTestAppForPackages(projectContext, _.extend(
     options,
-    { mobileServerUrl: utils.formatUrl(parsedMobileServerUrl) }));
+    {
+      mobileServerUrl: utils.formatUrl(parsedMobileServerUrl),
+      proxyPort: parsedServerUrl.port,
+      proxyHost: parsedServerUrl.host,
+    }
+  ));
 }
 
 // Returns the "local-test:*" package names for the given package names (or for
@@ -1755,7 +1756,8 @@ var runTestAppForPackages = function (projectContext, options) {
     var runAll = require('../runners/run-all.js');
     return runAll.run({
       projectContext: projectContext,
-      proxyPort: options.port,
+      proxyPort: options.proxyPort,
+      proxyHost: options.proxyHost,
       debugPort: options['debug-port'],
       disableOplog: options['disable-oplog'],
       settingsFile: options.settings,
@@ -1950,8 +1952,6 @@ main.registerCommand({
     throw new main.ShowUsage;
   }
 
-  config.printUniverseBanner();
-
   var username = options.add || options.remove;
 
   var conn = loggedInAccountsConnectionOrPrompt(
@@ -2007,6 +2007,10 @@ main.registerCommand({
     slow: { type: Boolean },
     galaxy: { type: Boolean },
     browserstack: { type: Boolean },
+    // Indicates whether these self-tests are running headless, e.g. in a
+    // continuous integration testing environment, where visual niceties
+    // like progress bars and spinners are unimportant.
+    headless: { type: Boolean },
     history: { type: Number },
     list: { type: Boolean },
     file: { type: String },
@@ -2086,6 +2090,12 @@ main.registerCommand({
   var clients = {
     browserstack: options.browserstack
   };
+
+  if (options.headless) {
+    // There's no point in spinning the spinner when we're running
+    // continuous integration tests.
+    Console.disableSpinner();
+  }
 
   return selftest.runTests({
     // filtering options