+++ tools/meteor-services/config.js
@@ -8,179 +8,54 @@ var tropohouse = require('../packaging/tropohouse.js');
 // deploying apps to the MDG free hosting sandbox, publishing packages,
 // getting an ssh access to a build farm. These functions need
 // configuration.
-//
-// The idea is that eventually, the `meteor` will take only one
-// configuration parameter, the "universe" it is talking to, which
-// defaults to "www.meteor.com". In a git checkout it can be set by
-// creating a file at the root of the checkout called "universe" that
-// contains the name of the universe you wish to use. Then, all other
-// needed configuration is derived from the universe name.
-//
-// We're not quite there yet though:
-// - When developing locally, you may need to set DISCOVERY_PORT (see
-//   getDiscoveryPort below)
-// - DEPLOY_HOSTNAME can still be set to override classic-style
-//   deploys
-// - The update/warehouse system hasn't been touched and still has its
-//   hardcoded URLs for now (update.meteor.com and
-//   warehouse.meteor.com). Really, it's debatable whether these
-//   should (necessarily) change when you change your universe name.
-
-var universe;
-var getUniverse = function () {
-  if (! universe) {
-    universe = "www.meteor.com";
-
-    if (files.inCheckout()) {
-      var p = files.pathJoin(files.getCurrentToolsDir(), 'universe');
-      if (files.exists(p)) {
-        universe = files.readFile(p, 'utf8').trim();
-      }
-    }
-  }
-
-  return universe;
-};
-
-var isLocalUniverse = function () {
-  return !! getUniverse().match(/^localhost(:([\d]+))?$/);
-};
-
-var localhostOffset = function (portOffset) {
-  var match = getUniverse().match(/^localhost(:([\d]+))?$/);
-  if (! match) {
-    throw new Error("not a local universe?");
-  }
-  return "localhost:" + (parseInt(match[2] || "80") + portOffset);
-};
-
-var getAuthServiceHost = function () {
-  if (! isLocalUniverse()) {
-    return universe;
-  } else {
-    // Special case for local development. Point
-    // $METEOR_CHECKOUT/universe at the place where you are running
-    // frontpage (eg, localhost:3000), and run the accounts server ten
-    // port numbers higher. Like so:
-    //   cd meteor-accounts
-    //   ROOT_URL=http://localhost:3010/auth curmeteor -p 3010
-    return localhostOffset(10);
-  }
-};
-
-// Given a hostname, add "http://" or "https://" as
-// appropriate. (localhost gets http; anything else is always https.)
-var addScheme = function (host) {
-  if (host.match(/^localhost(:\d+)?$/)) {
-    return "http://" + host;
-  } else {
-    return "https://" + host;
-  }
-};
 
 var config = exports;
 _.extend(exports, {
-  // True if this the production universe (www.meteor.com)
-  isProduction: function () {
-    return getUniverse() === "www.meteor.com";
-  },
-
-  // The current universe name. Should be used for cosmetic purposes
-  // only (displaying to the user). If you want to programmatically
-  // derive configuration from it, add a new method to this file.
-  getUniverse: function () {
-    return getUniverse();
-  },
-
-  // Base URL for Meteor Accounts OAuth services, typically
-  // "https://www.meteor.com/oauth2". Endpoints include /authorize and
-  // /token.
+  // Base URL for Meteor Accounts OAuth services. Endpoints include /authorize
+  // and /token.
   getOauthUrl: function () {
-    return addScheme(getAuthServiceHost()) + "/oauth2";
+    return "https://www.meteor.com/oauth2";
   },
 
-  // Base URL for Meteor Accounts API, typically
-  // "https://www.meteor.com/api/v1". Endpoints include '/login' and
+  // Base URL for Meteor Accounts API. Endpoints include '/login' and
   // '/logoutById'.
   getAccountsApiUrl: function () {
-    return addScheme(getAuthServiceHost()) + "/api/v1";
+    return "https://www.meteor.com/api/v1";
   },
 
-  // URL for the DDP interface to Meteor Accounts, typically
-  // "https://www.meteor.com/auth". (Really should be a ddp:// URL --
-  // we'll get there soon enough.)
+  // URL for the DDP interface to Meteor Accounts.
   getAuthDDPUrl: function () {
-    return addScheme(getAuthServiceHost()) + "/auth";
+    return "https://www.meteor.com/auth";
   },
 
   // URL for the DDP interface to the meteor build farm, typically
   // "https://build.meteor.com".
   getBuildFarmUrl: function () {
-    if (process.env.METEOR_BUILD_FARM_URL) {
-      return process.env.METEOR_BUILD_FARM_URL;
-    }
-    var host = config.getBuildFarmDomain();
-
-    return addScheme(host);
+    return process.env.METEOR_BUILD_FARM_URL || "https://build.meteor.com";
   },
 
   getBuildFarmDomain: function () {
-    if (process.env.METEOR_BUILD_FARM_URL) {
-      var parsed = url.parse(process.env.METEOR_BUILD_FARM_URL);
-      return parsed.host;
-    } else {
-      return getUniverse().replace(/^www\./, 'build.');
-    }
+    return url.parse(config.getBuildFarmUrl()).host;
   },
 
   // URL for the DDP interface to the package server, typically
-  // "https://packages.meteor.com". (Really should be a ddp:// URL --
-  // we'll get there soon enough.)
-  //
-  // When running everything locally, run the package server at the
-  // base universe port number (that is, the Meteor Accounts port
-  // number) plus 20.
+  // "https://packages.meteor.com".
   getPackageServerUrl: function () {
-    if (process.env.METEOR_PACKAGE_SERVER_URL) {
-      return process.env.METEOR_PACKAGE_SERVER_URL;
-    }
-    var host = config.getPackageServerDomain();
-
-    return addScheme(host);
+    return process.env.METEOR_PACKAGE_SERVER_URL ||
+      "https://packages.meteor.com";
   },
 
   getPackageServerDomain: function () {
-    if (isLocalUniverse()) {
-      return localhostOffset(20);
-    } else {
-      if (process.env.METEOR_PACKAGE_SERVER_URL) {
-        var parsed = url.parse(process.env.METEOR_PACKAGE_SERVER_URL);
-        return parsed.host;
-      } else {
-        return getUniverse().replace(/^www\./, 'packages.');
-      }
-    }
+    return url.parse(config.getPackageServerUrl()).host;
   },
 
   getPackageStatsServerUrl: function () {
-    if (process.env.METEOR_PACKAGE_STATS_SERVER_URL) {
-      return process.env.METEOR_PACKAGE_STATS_SERVER_URL;
-    }
-
-    var host = config.getPackageStatsServerDomain();
-    return addScheme(host);
+    return process.env.METEOR_PACKAGE_STATS_SERVER_URL ||
+      "https://activity.meteor.com";
   },
 
   getPackageStatsServerDomain: function () {
-    if (process.env.METEOR_PACKAGE_STATS_SERVER_URL) {
-      return url.parse(process.env.METEOR_PACKAGE_STATS_SERVER_URL).hostname;
-    }
-
-    if (isLocalUniverse()) {
-      return localhostOffset(30);
-    } else {
-      return getUniverse().replace(/^www\./, 'activity.');
-    }
+    return url.parse(config.getPackageStatsServerUrl()).host;
   },
 
   // Note: this is NOT guaranteed to return a distinct prefix for every
@@ -246,56 +121,7 @@ _.extend(exports, {
   // use. This is used as a key for storing your Meteor Accounts
   // login token.
   getAccountsDomain: function () {
-    return getUniverse();
-  },
-
-  getDeployHostname: function () {
-    return process.env.DEPLOY_HOSTNAME || "meteor.com";
-  },
-
-  getFullAppName: function(appName) {
-    var domain = process.env.DEPLOY_DOMAIN || this.getDeployHostname();
-    if (appName.indexOf(".") === -1) {
-      return appName + "." + domain;
-    }
-    return appName; 
-  },
-
-  // Deploy URL for MDG free hosting, eg 'https://deploy.meteor.com'.
-  getDeployUrl: function () {
-    var host;
-
-    // Support the old DEPLOY_HOSTNAME environment variable for a
-    // while longer. Soon, let's remove this in favor of the universe
-    // scheme.
-    if (process.env.DEPLOY_HOSTNAME) {
-      host = process.env.DEPLOY_HOSTNAME;
-      if (host.match(/^http/)) {
-        // allow it to contain a URL scheme
-        return host;
-      }
-    } else {
-      // Otherwise, base it on the universe.
-      if (isLocalUniverse()) {
-        throw new Error("local development of deploy server not supported");
-      } else {
-        host = getUniverse().replace(/^www\./, 'deploy.');
-      }
-    }
-
-    return addScheme(host);
-  },
-
-  // URL from which the update manifest may be fetched, eg
-  // 'https://update.meteor.com/manifest.json'
-  getUpdateManifestUrl: function () {
-    if (isLocalUniverse()) {
-      // localhost can't run the manifest server
-      u = "www.meteor.com";
-    }
-    var host = getUniverse().replace(/^www\./, 'update.');
-
-    return addScheme(host) + "/manifest.json";
+    return "www.meteor.com";
   },
 
   // Path to file that contains our credentials for any services that
@@ -305,32 +131,5 @@ _.extend(exports, {
     // METEOR_SESSION_FILE is for automated testing purposes only.
     return process.env.METEOR_SESSION_FILE ||
       files.pathJoin(files.getHomeDir(), '.meteorsession');
-  },
-
-  // Port to use when querying URLs for the deploy server that backs
-  // them, and for querying oauth clients for their oauth information
-  // (so we can log into them).
-  //
-  // In production this should always be 443 (we *must*
-  // cryptographically authenticate the server answering the query),
-  // but this can be inconvenient for local development since 443 is a
-  // privileged port, so you can set DISCOVERY_PORT to override. (A
-  // better solution would probably be to spin up a local VM.)
-  getDiscoveryPort: function () {
-    if (process.env.DISCOVERY_PORT) {
-      return parseInt(process.env.DISCOVERY_PORT);
-    } else {
-      return 443;
-    }
-  },
-
-  // It's easy to forget that you're in an alternate universe (and
-  // that that is the reason you're not seeing your deploys). If not
-  // in production mode, print a quick hint about the universe you're
-  // in.
-  printUniverseBanner: function () {
-    if (! config.isProduction()) {
-      process.stderr.write('[Universe: ' + config.getUniverse() + ']\n');
-    }
   }
 });