+++ tools/tool-testing/test-utils.js
@@ -48,11 +48,6 @@ exports.logout = function (s) {
   run.expectExit(0);
 };
 
-exports.getUserId = function (s) {
-  var data = JSON.parse(s.readSessionFile());
-  return data.sessions[config.getUniverse()].userId;
-};
-
 var registrationUrlRegexp =
       /https:\/\/www\.meteor\.com\/setPassword\?([a-zA-Z0-9\+\/]+)/;
 exports.registrationUrlRegexp = registrationUrlRegexp;